Fix deadlock on dispose
The app has an issue where it doesn't close properly. It's very common
on Linux but no so much on Windows, probably due to the differences in
synchronizer primitives.

Pausing the app with the debugger while it's stuck reveals that it's
waiting in `_pending Events.Reader.ReadAsync`. Originally, the code
Connection `Dispose` method would just try and complete the writer of
the channel which would result in `OperationCanceledException` being
thrown when the reader then tries to read from the channel.

However, I don't think that ever happens. `Reader.ReadAsync` just seems
to be stuck on some lock and since we didn't provide a CancellationToken
it's stuck forever.

This PR adds a cancellation token so that the thread and the entire app
can properly exist.